requirements.txt
a2a-sdk[http-server]
httpx
uvicorn
starlette
requests